VillainAI/non openrouter/requirements.txt
flask
python-dotenv
requests
google-generativeai
mistralai 
cohere